+++ sArena_Init.lua
@@ -8,7 +8,10 @@ sArenaMixin.isMoP = gameVersion:match("^5%.")
 sArenaMixin.isWrath = gameVersion:match("^3%.")
 sArenaMixin.isTBC = gameVersion:match("^2%.")
 
-sArenaMixin.addonTitle = "|T135884:13:13|t sArena |cffff8000Reloaded|r " .. (C_AddOns.GetAddOnMetadata("sArena_Reloaded", "Version") or "")
+sArenaMixin.addonName = "|T135884:13:13|t sArena |cffff8000Reloaded|r"
+sArenaMixin.addonTitle = sArenaMixin.addonName.. " " .. (C_AddOns.GetAddOnMetadata("sArena_Reloaded", "Version") or "")
+sArenaMixin.popupHeader = "\n"..sArenaMixin.addonName.."\n\n"
+
 sArenaMixin.layouts = {}
 sArenaMixin.defaultSettings = {
     profile = {
@@ -49,5 +52,9 @@ sArenaMixin.defaultSettings = {
         },
         layoutSettings = {},
         invertClassIconCooldown = true,
+        clickAttributes = {
+            ["Left"] = { button = "1", action = "target" },
+            ["Right"] = { button = "2", action = "focus" },
+        },
     }
 }