+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# sArena Reloaded 2.4.4
+## New
+- Click Actions (Global). You can now add new and customize existing Click Actions like Left Click to Target and Right Click to Focus. For example Shift+RightClick to use a macro.
+- Midnight only: New castbar setting "Highlight Casts on Me" that puts a bright border around the castbar if the spell is being casted on you. (Layout -> Cast Bar)
+## Tweak
+- Fix Castbar TargetText/ID not getting proper font sometimes.
+- Castbar color tweaks to be more consistent.
+- Midnight: Fix castbar interrupt detection so now castbars will immediately hide again when a cast is over except for when they get interrupted and it will say who interrupted it and fade out slowly.
+## Bugfix
+- Midnight: Fix percent display on manabars being stuck at 0-1 instead of 0-100.
+- Midnight: Fix DR Black Border setting being white.
+
 # sArena Reloaded 2.4.3b
 ## New
 - Color Trinket: Keep Original Texture; Instead of replacing the texture entirely with a solid color keep the original texture but tinted in a color.